requirements.txt
pdfplumber==0.11.2
transformers==4.42.4
torch==2.3.1
nltk==3.8.1
python-dotenv==1.0.1
semantic_router==0.0.54
pinecone-client==5.0.0
pinecone==4.0.0
groq==0.9.0
tqdm==4.66.4
datasets==2.20.0
rouge_score==0.1.2
pandas==2.2.0
